Create binary_models.RDS
This is a temporary fix, as I noticed a coding issue in the neighbor binary outputs (namely that it was a total and not a 0/1 output). Due to time constraints, I recoded those columns in post and re-ran the binary models. will need to fix cleaning code in the future